+++ src/Catty/FormulaElement.m
@@ -203,7 +203,7 @@
             while (left < 0) {
                 left += right;
             }
-            result = fmodf(left, right);
+            result = fmod(left, right);
             break;
         }
         case ARCSIN: {
