+++ src/Catty/SetVolumeToBrick.m
@@ -40,7 +40,7 @@
     return [SKAction runBlock:^{
         NSDebug(@"Performing: %@", self.description);
         double volume = [self.volume interpretDoubleForSprite:self.object];
-        [[AudioManager sharedAudioManager] setVolumeToPercent:volume forKey:self.object.name];
+        [[AudioManager sharedAudioManager] setVolumeToPercent:(CGFloat)volume forKey:self.object.name];
         
     }];
 }
