+++ src/Catty/TurnLeftBrick.m
@@ -51,7 +51,7 @@
         else if (newRad <= (- 2*M_PI)) {
             newRad += 2*M_PI;
         }
-        self.object.zRotation = newRad;
+        self.object.zRotation = (CGFloat)newRad;
     };
 }
